+++ src/synapsis/core/__init__.py
@@ -1,3 +1,4 @@
+from .narg import Narg, none
 from .utils import Utils
 from .hooks import Hooks
 from .synapsis import Synapsis