+++ tests/synapsis/synapse/test_synapse_permission.py
@@ -1,8 +1,8 @@
 import pytest
 from synapsis.synapse import SynapsePermission
 from synapsis.core.exceptions import NotFoundError
-from synapsis.core import Utils
-from synapsis.synapse.synapse_permission import SynapsePermissions
+from synapsis.core import Utils, none
+import itertools
 
 entity_permission_codes = [
     'NO_PERMISSION',
@@ -21,25 +21,25 @@ team_permission_codes = [
 all_permission_codes = set(entity_permission_codes + team_permission_codes)
 
 
-def test_contexts():
+def test_collections():
     entity_permissions = SynapsePermission.ENTITY_PERMISSIONS
     assert len(entity_permissions) == len(entity_permission_codes)
     for code in entity_permission_codes:
-        p = Utils.find(entity_permissions, lambda ep: ep.code == code, None)
+        p = Utils.find(entity_permissions, lambda ep: ep.code == code)
         assert p is not None
         assert p.code == code
 
     team_permissions = SynapsePermission.TEAM_PERMISSIONS
     assert len(team_permissions) == len(team_permission_codes)
     for code in team_permission_codes:
-        p = Utils.find(team_permissions, lambda ep: ep.code == code, None)
+        p = Utils.find(team_permissions, lambda ep: ep.code == code)
         assert p is not None
         assert p.code == code
 
     all_permissions = SynapsePermission.ALL
     assert len(all_permissions) == len(all_permission_codes)
     for code in all_permission_codes:
-        p = Utils.find(all_permissions, lambda ep: ep.code == code, None)
+        p = Utils.find(all_permissions, lambda ep: ep.code == code)
         assert p is not None
         assert p.code == code
 
@@ -52,56 +52,70 @@ def test_cached_attributes():
         assert perm1 == perm2
 
 
-def test_find_by():
+def test_get():
+    invalid = SynapsePermission(code='INVALID', name='INVALID', access_types=['INVALID'])
     for perm in SynapsePermission.ALL:
-        assert SynapsePermission.find_by(code=perm.code) == perm
-        assert SynapsePermission.find_by(code=perm.code.lower()) == perm
+        for arg in [perm, perm.code.upper(), perm.code.lower(), perm.access_types] + \
+                   [list([a.lower() for a in perm.access_types])] + \
+                   [list([a.upper() for a in perm.access_types])]:
+            assert SynapsePermission.get(arg) == perm
 
-        assert SynapsePermission.find_by(access_types=perm.access_types) == perm
-        assert SynapsePermission.find_by(access_types=[a.lower() for a in perm.access_types]) == perm
-
-        assert SynapsePermission.find_by(code=perm.code, access_types=perm.access_types) == perm
-        assert SynapsePermission.find_by(code='NOPE', access_types=perm.access_types) is None
-        assert SynapsePermission.find_by(code=perm.code, access_types=['NOPE']) is None
-
-        with pytest.raises(ValueError) as ex:
-            SynapsePermission.find_by()
-        assert 'code or access_types are required.' in str(ex)
-
-        assert SynapsePermission.find_by(code='NOPE', raises=False) is None
-
-        with pytest.raises(NotFoundError) as ex:
-            SynapsePermission.find_by(code='NOPE', raises=True)
-        assert 'Could not find SynapsePermission' in str(ex)
+    for arg, default, expected in [
+        (None, None, None),
+        (None, SynapsePermission.NO_PERMISSION, SynapsePermission.NO_PERMISSION),
+        (None, SynapsePermission.ADMIN, SynapsePermission.ADMIN),
+        (invalid, SynapsePermission.NO_PERMISSION, SynapsePermission.NO_PERMISSION),
+        (invalid, None, None),
+        (invalid.code, None, None),
+        (invalid.access_types, None, None),
+        (invalid, none, NotFoundError)
+    ]:
+        if expected == NotFoundError:
+            with pytest.raises(expected):
+                SynapsePermission.get(arg, default)
+        else:
+            assert SynapsePermission.get(arg, default) == expected
 
 
 def test_are_equal():
     invalid = SynapsePermission(code='INVALID', name='INVALID', access_types=['INVALID'])
+    assert SynapsePermission.get(invalid, None) is None
+    assert SynapsePermission.are_equal(None, None) is False
+
     for perm in SynapsePermission.ALL:
         same = SynapsePermission(code=perm.code, name=perm.name, access_types=perm.access_types)
-        assert perm != same
-        assert perm.access_types == same.access_types
+        other = Utils.first(SynapsePermission.ALL, lambda p: p.code != perm.code)
+        assert same != perm
+        assert same.code == perm.code
+        assert same.access_types == perm.access_types
         assert perm.equals(same)
+        assert same.equals(perm)
+        assert other.code != perm.code
+        assert other.access_types != perm.access_types
+        assert SynapsePermission.are_equal(perm, other) is False
 
         # True
-        for args in [
-            (perm, same),
-            (perm.access_types, same),
-            (perm, same.access_types),
-            (perm.access_types, same.access_types)
-        ]:
-            assert SynapsePermission.are_equal(*args)
-            assert SynapsePermission.are_equal(*args)
+        permutations = list(itertools.permutations(
+            [perm, perm.code, perm.access_types, same, same.code, same.access_types], r=2))
+        for a, b in permutations:
+            assert SynapsePermission.are_equal(a, b)
+            if isinstance(a, SynapsePermission):
+                assert a.equals(b)
+            if isinstance(b, SynapsePermission):
+                assert b.equals(a)
 
         # False
-        for arg in [invalid, None]:
-            assert perm.equals(arg) is False
+        products = list(itertools.product(
+            [perm, perm.code, perm.access_types, invalid, invalid.code, invalid.access_types],
+            [other, other.code, other.access_types, '', ' ', 'NOPE']
+        ))
+        for a, b in products:
+            a_perm = SynapsePermission.get(a, None)
+            b_perm = SynapsePermission.get(b, None)
+            expected = a_perm == SynapsePermission.NO_PERMISSION and b_perm == SynapsePermission.NO_PERMISSION
 
-        for args in [
-            (None, same.access_types),
-            (perm.access_types, None),
-            (None, None),
-            (perm, invalid)
-        ]:
-            assert SynapsePermission.are_equal(*args) is False
-            assert SynapsePermission.are_equal(*args) is False
+            assert SynapsePermission.are_equal(a, b) is expected
+            if isinstance(a, SynapsePermission):
+                assert a.equals(b) is expected
+            if isinstance(b, SynapsePermission):
+                assert b.equals(a) is expected